dashboard/requirements.txt
flask
happybase
xgboost
numpy
joblib
google-generativeai
newsapi-python
pymongo
pandas